+++ internal/provider/platform_resource_schema_gcp.go
@@ -1,8 +1,11 @@
 package provider
 
 import (
+	"github.com/hashicorp/terraform-plugin-framework-validators/objectvalidator"
+	"github.com/hashicorp/terraform-plugin-framework/path"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
+	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 )
 
 func gcpPlatformSchema() schema.Attribute {
@@ -82,7 +85,7 @@ func gcpReplicationConfigSchema() schema.Attribute {
 
 func gcpServiceAccountConfigSchema() schema.Attribute {
 	return schema.SingleNestedAttribute{
-		MarkdownDescription: "Service account configuration. Either credential or workload_identity must be provided.",
+		MarkdownDescription: "Service account configuration. Exactly one of credential or workload_identity must be provided.",
 		Required:            true,
 		Attributes: map[string]schema.Attribute{
 			"type": schema.StringAttribute{
@@ -94,6 +97,15 @@ func gcpServiceAccountConfigSchema() schema.Attribute {
 			"workload_identity": schema.SingleNestedAttribute{
 				MarkdownDescription: "Workload identity configuration.",
 				Optional:            true,
+				Validators: []validator.Object{
+					// This will result in an unclear path in the error message for credential because
+					// it traverses up to the parent and back down.
+					// see https://github.com/hashicorp/terraform-plugin-framework-validators/issues/274
+					objectvalidator.ExactlyOneOf(
+						path.MatchRelative(),
+						path.MatchRelative().AtParent().AtName("credential"),
+					),
+				},
 				Attributes: map[string]schema.Attribute{
 					"audience": schema.StringAttribute{
 						MarkdownDescription: "The audience associated with your workload identity pool provider.",